get monthly and yearly productions fix

+++ Backend/DTO/Responses/MonthlyEnergyProductionResponse.cs
@@ -1,5 +1,5 @@
 ﻿namespace Backend.DTO.Responses
 {
-    public record MonthlyProduction(int month, int year, int production);
+    public record MonthlyProduction(int month, int year, float production);
     public record MonthlyEnergyProductionsResponse(string serialNumber, List<MonthlyProduction> monthlyProductions);
 }
