+++ Backend/DTO/Responses/YearlyEnergyProductionsResponse.cs
@@ -1,5 +1,5 @@
 namespace Backend.DTO.Responses
 {
-    public record YearlyProduction(int year, int production);
+    public record YearlyProduction(int year, float production);
     public record YearlyEnergyProductionsResponse(string serialNumber, List<YearlyProduction> monthlyProductions);
 }
